@@ -1,53 +1,30 @@
-require('dotenv').config(); // Load environment variables from .env file
-const express = require("express");
-const cors = require('cors');
-const app = express();
-const { Sequelize, DataTypes } = require('@sequelize/core');
-const { PostgresDialect } = require('@sequelize/postgres');
+import dotenv from 'dotenv';
+import express from 'express';
+import userRoute from './routes/user.route.js';
+import mongoose from 'mongoose';
+import cors from 'cors';
 
+dotenv.config()
+
+//creates the server
+const app = express(); 
+
+//allows us to accept JSON data in req.body
+app.use(express.json());
+
+//allows to server to interact with the front end
 app.use(cors());
 
-// Set up Sequelize using the render DATABASE_URL stored in  .env
-const sequelize = new Sequelize({
-    dialect: PostgresDialect,
-    url: process.env.DATABASE_URL, // Use the DATABASE_URL from .env
-    ssl: true,
-    clientMinMessages: 'notice',
-});
-  
-// Database connection tester
-sequelize.authenticate()
-  .then(() => console.log('Database connected successfully...'))
-  .catch(err => console.error('Unable to connect to the database:', err));
+//application functions
+app.use('',userRoute);
 
-// Db simple model definition
-const User = sequelize.define('User', {
-    username: { type: DataTypes.STRING, allowNull: false }, 
-    email: { type: DataTypes.STRING, allowNull: false, unique: true }, 
-});
-  
-// Sync the database (create tables)
-sequelize.sync({ force: true });
-
-// API Route for Hello World message
-app.get("/api/home", (req, res) => {
-    res.json({ message: "Hello World!" });
-});
-  
-// API Route to fetch users from the database
-app.get('/api/users', async (req, res) => {
-    try {
-         // Fetch all users from the database
-      const users = await User.findAll();
-      res.json(users);
-    } catch (err) {
-      console.error('Error fetching users:', err);
-      res.status(500).json({ error: 'Failed to fetch users' });
-    }
-});
-  
-// Server Starter
-const port = process.env.PORT || 8080;
-app.listen(port, () => {
-    console.log(`Server started on port ${port}`);
+//creates server on port 8080 and connects to mongodb database
+app.listen(process.env.PORT || 8080, '0.0.0.0', () => {
+    mongoose.connect(process.env.mongodb_URI)
+    .then((result) => {
+        console.log('connected to Mongodb');
+    }).catch((err) => {
+        console.error(err);
+    });
+    console.log('Server started at http://localhost:' + process.env.PORT);
 });
